Exorcise Storybook and De-cruft
Hopefully will speed up test runner on GitHub
