+++ Dockerfile.dev
@@ -30,7 +30,4 @@ WORKDIR /app
 # run empricia alone with logging
 #CMD ["empirica", "--log.json", "--tajriba.log.json"]
 
-# run empirica dev mode and storybook together
-# CMD ["bash", "-c", "(cd /app/client; npm run storybook) & (cd /app; empirica)"] 
-
 # & (cd /app/client; npx cypress open)
