+++ client/package.json
@@ -14,12 +14,6 @@
   },
   "devDependencies": {
     "@babel/core": "^7.17.8",
-    "@storybook/addon-actions": "^6.4.19",
-    "@storybook/addon-essentials": "^6.4.19",
-    "@storybook/addon-interactions": "^6.4.19",
-    "@storybook/addon-links": "^6.4.19",
-    "@storybook/react": "^6.4.19",
-    "@storybook/testing-library": "^0.0.9",
     "@types/react": "17.0.18",
     "@types/react-dom": "17.0.9",
     "@vitejs/plugin-react-refresh": "1.3.1",
@@ -35,9 +29,7 @@
   "scripts": {
     "build": "NODE_ENV=\"production\" vite build",
     "dev": "vite",
-    "serve": "vite preview",
-    "storybook": "start-storybook -p 6006",
-    "build-storybook": "build-storybook"
+    "serve": "vite preview"
   },
   "version": "0.0.0",
   "volta": {
