--- client/src/components/Topic.test.jsx
@@ -1,1 +1,0 @@
-// Tests go here.
