--- client/src/intro-exit/Surveys/ExampleExitSurvey.jsx
@@ -1,15 +1,0 @@
-import { usePlayer } from "@empirica/player";
-import React, { useState } from "react";
-import { render } from "react-dom";
-import { Alert } from "../../components/Alert";
-import { Button } from "../../components/Button";
-import SurveyWrapper from "../../components/SurveyWrapper";
-import surveyJSON from './ExampleSurvey.json';
-
-export default function ExampleExitSurvey({ next }) {
-
-    // TODO: empirica callbacks on survey completion
-    return(
-        <SurveyWrapper surveyJson={surveyJSON} next={next} />
-    )
-}
