--- client/src/intro-exit/Surveys/team_viability.json
@@ -1,74 +1,0 @@
-{
-  "title": "Team Viability Assessment",
-  "logoPosition": "right",
-  "focusFirstQuestionAutomatic": false,
-  "pages": [
-   {
-    "name": "page1",
-    "elements": [
-     {
-      "type": "matrix",
-      "name": "team-viability",
-      "minWidth": "500px",
-      "maxWidth": "Initial",
-      "title": "Please select the option that most closely reflects how much you agree with the following statements.",
-      "titleLocation": "top",
-      "hideNumber": true,
-      "isRequired": true,
-      "columns": [
-       {
-        "value": "1",
-        "text": "Strongly disagree"
-       },
-       {
-        "value": "2",
-        "text": "Disagree"
-       },
-       {
-        "value": "3",
-        "text": "Neither agree nor disagree"
-       },
-       {
-        "value": "4",
-        "text": "Agree"
-       },
-       {
-        "value": "5",
-        "text": "Strongly Agree"
-       }
-      ],
-      "rows": [
-       {
-        "value": "capable unit",
-        "text": "This team is capable of working together as a unit."
-       },
-       {
-        "value": "future success",
-        "text": "This team would work well together in the future."
-       },
-       {
-        "value": "falling apart",
-        "text": "As a team, this work group shows signs of falling apart."
-       },
-       {
-        "value": "welcome reunion",
-        "text": "Most of the members of this team would welcome the opportunity to work as a group again in the future."
-       },
-       {
-        "value": "persist despite obstacles",
-        "text": "This team has what it takes to persist in the face of obstacles."
-       },
-       {
-        "value": "succeed dispite dislike",
-        "text": "This team will still be successful even if the members do not like one another."
-       }
-      ]
-     }
-    ],
-    "questionTitleLocation": "top"
-   }
-  ],
-  "showPageTitles": false,
-  "questionsOnPageMode": "singlePage",
-  "widthMode": "responsive"
- }
