+++ cypress/e2e/01_Normal_Paths.js
@@ -4,7 +4,7 @@
 
 import dayjs from "dayjs";
 
-describe("Normal Paths: Control", { retries: { runMode: 2, openMode: 0 } }, () => {
+describe("Normal paths: Control", { retries: { runMode: 2, openMode: 0 } }, () => {
     let start; 
     let end; 
     let difference; 
@@ -118,7 +118,7 @@ describe("Normal Paths: Control", { retries: { runMode: 2, openMode: 0 } }, () =
             end = dayjs();
             difference = end.diff(start)
             payment = ((difference / 3600000) * 15)
-            cy.contains("calculating", { timeout: 15000 }).should('not.exist');
+            cy.contains("calculating", { timeout: 20000 }).should('not.exist');
             cy.get(`[data-test="dollarsOwed"]`).invoke('text').then($value => cy.log(`Observed payment ${$value}`))
             cy.get(`[data-test="dollarsOwed"]`).invoke('text').then(parseFloat).should('be.closeTo', payment, .02)
       });
