+++ cypress/e2e/05_Mobile_Check.js
@@ -2,8 +2,7 @@
 // This test aims to test what a user will encounter if they
 // attempt to complete the experiment on a mobile device
 
-describe("mobile_check", () => {
-  const playerKey = "test_" + Math.floor(Math.random() * 1e13);
+describe("Disallow mobile connections", () => {
 
   before(() => {
     cy.empiricaClearBatches();
@@ -24,6 +23,7 @@ describe("mobile_check", () => {
   });
 
   it("redirects to mobile error screen", () => {
+    const playerKey = "test_" + Math.floor(Math.random() * 1e13);
     cy.empiricaLoginPlayer(playerKey);
     cy.viewport("iphone-x");
