+++ docker-compose-dev.yml
@@ -16,23 +16,3 @@ services:
     command:
       - empirica
 
-
-  storybook:
-    platform: linux/amd64  # needed when developing on apple silicon
-    build:
-      context: .
-      dockerfile: Dockerfile.dev
-    ports:
-      - "6006:6006"  # storybook
-    volumes:
-      - .:/app
-      - /app/client/node_modules  # should ensure we don't clobber the container modules with the local.
-      - /app/server/node_modules
-    working_dir: /app/client
-    command:
-      - npm
-      - run
-      - storybook
-
-
-
